+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed 
 
 - Broken import in ftp filesystem module.
+- The intorods.filesys.sync.sync method uses `<fs>.mkdir(path, parents=True)`, added parents kw to mkdir for all filesystems.
 
 ## [0.0.7] - 2023-06-07